backend/requirements.txt
# FastAPI and server
fastapi==0.110.0
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# Database
sqlalchemy[asyncio]==2.0.27
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9

# Redis and Caching
redis==5.0.1
hiredis==2.3.2

# Celery for background jobs
celery==5.3.6
flower==2.0.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.9
authlib==1.3.0

# AI/ML - Multi-Provider Support
litellm>=1.40.0
langchain>=0.2.0
langchain-openai>=0.1.0
openai>=1.30.0
anthropic>=0.25.0
google-generativeai>=0.5.0
tiktoken>=0.7.0

# Vector Database
qdrant-client==1.7.3

# File Processing & Storage
google-cloud-storage==2.16.0
pandas==2.2.1
openpyxl==3.1.2

# HTTP Clients
httpx>=0.28.1
aiohttp>=3.10.0

# Jira Integration
jira==3.6.0

# Data Validation
pydantic==2.6.1
pydantic-settings==2.1.0
email-validator==2.1.0

# Encryption
cryptography==42.0.2

# Utilities
python-dotenv==1.0.1
orjson==3.9.15
pytz==2024.1

# Testing
pytest==7.4.4
pytest-asyncio==0.23.4
pytest-cov==4.1.0
pytest-mock==3.12.0
httpx>=0.28.1
faker==23.1.0

# Code Quality
black==24.1.1
flake8==7.0.0
mypy==1.8.0
isort==5.13.2
pylint==3.0.3

# Monitoring
sentry-sdk[fastapi]==1.40.0
google-cloud-aiplatform>=1.130.0
